fix: restore platform-dependent u64 casts in disk space calc
The previous clippy --fix removed "as u64" casts on statvfs fields that
are u64 on Linux but u32 on macOS. Removing them broke the macOS build
(cannot multiply u32 by u64). The cast is required for cross-platform
correctness; the "unnecessary cast" lint is a Linux-only false positive.

+++ src/app.rs
@@ -874,8 +874,8 @@ fn get_disk_space(path: &std::path::Path) -> Option<(u64, u64)> {
         let mut stat = MaybeUninit::<libc::statvfs>::uninit();
         if libc::statvfs(c_path.as_ptr(), stat.as_mut_ptr()) == 0 {
             let stat = stat.assume_init();
-            let total = stat.f_blocks * stat.f_frsize;
-            let free = stat.f_bavail * stat.f_frsize;
+            let total = stat.f_blocks as u64 * stat.f_frsize;
+            let free = stat.f_bavail as u64 * stat.f_frsize;
             Some((total, free))
         } else {
             None
